requirements.txt
cachetools==5.3.0
Flask==2.2.2
Jinja2==3.1.2
nested_lookup==0.2.25
Pillow==9.5.0
requests==2.22.0
telegrambots==0.0.13rc0
